requirements.txt
httpx==0.27.2
pydantic==2.9.2
rich==13.9.4
python-dateutil==2.9.0.post0
